add yaml parser, move files around
