+++ package.json
@@ -25,5 +25,8 @@
     "tslib": "^2.3.1",
     "typescript": "^4.4.3"
   },
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "js-yaml": "^4.1.0"
+  }
 }
