+++ server/yaml.json
@@ -1,0 +1,49 @@
+[
+  {
+    "apiVersion": "apps/v1",
+    "kind": "Deployment",
+    "metadata": {
+      "name": "mafia-backend",
+      "namespace": "mafia",
+      "labels": {
+        "name": "mafia-backend",
+        "app": "mafia"
+      }
+    },
+    "spec": {
+      "replicas": 1,
+      "selector": {
+        "matchLabels": {
+          "app": "mafia"
+        }
+      },
+      "template": {
+        "metadata": {
+          "labels": {
+            "name": "mafia-backend",
+            "app": "mafia"
+          }
+        },
+        "spec": {
+          "containers": [
+            {
+              "name": "mafia-backend-container",
+              "image": "taleodor/mafia-express@sha256:8328cd7de911367ef2a52a60f8696720e89c152c9a437983c7eccf411b0ede75",
+              "env": [
+                {
+                  "name": "REDIS_HOST",
+                  "value": "redis-service.mafia"
+                }
+              ],
+              "ports": [
+                {
+                  "containerPort": 3000
+                }
+              ]
+            }
+          ]
+        }
+      }
+    }
+  }
+]
