--- src/node_types/kObject.ts
@@ -1,4 +1,0 @@
-export interface kObject {
-  label: string;
-  getLabel: () => string;
-}
