--- src/node_types/kStatefulSet.ts
@@ -1,16 +1,0 @@
-import type {kObject} from './kObject';
-
-export class kStatefulSet implements kObject {
-  label: string;
-  replicas: number;
-  
-  constructor(label: string, replicas: number)
-  {
-    this.label = label;
-    this.replicas = replicas;
-  }
-  getLabel(): string
-  {
-    return this.label;
-  }
-}
